+++ src/main/java/racingcar/domain/Car.java
@@ -1,25 +1,11 @@
 package racingcar.domain;
 
 public class Car {
-    private String carName;
+    private CarName carName;
     private int movingDistance;
 
     public Car(String carName) {
-        validateCarNameLengthExceed(carName);
-        validateCarNameLengthZero(carName);
-        this.carName = carName;
+        this.carName = new CarName(carName);
         this.movingDistance = 0;
-    }
-
-    private void validateCarNameLengthExceed(String carName) {
-        if (carName.length() > 5) {
-            throw new IllegalArgumentException("자동차의 이름은 5자를 넘어갈 수 없습니다.");
-        }
-    }
-
-    private void validateCarNameLengthZero(String carName) {
-        if (carName.isEmpty()) {
-            throw new IllegalArgumentException("자동차의 이름은 공백이 될 수 없습니다.");
-        }
     }
 }
